separate db config
